oss: move server_config into oss namespace
