+++ src/main.cpp
@@ -1,5 +1,5 @@
 #include "ui/login_component.hpp"
-#include "server_config.hpp"
+#include "oss/server_config.hpp"
 
 import nlohmann.json;
 import ftxui;
@@ -8,7 +8,7 @@ int main()
 {
     namespace ui = ftxui;
 
-    server_config config{"default_name", "https://localhost"};
+    oss::server_config config{"default_name", "https://localhost"};
     auto screen{ui::ScreenInteractive::Fullscreen()};
 
     login_component login{screen, config};
